Update before code restructuring ================================
After this commit, the configuration of WFCs and WFSs will be different,
and specific WFS details will be split off to modules. This means for
example that all SH related configuration of the WFS, which is *not* generic
for a WFS, will be split off to the sh module.

This requires some recoding of the main framework and the modules, therefore
this is a last save before this operation. Also, the modules will be using
less globals unless it is absolutely necessary. Also take a look at the
itifg, daq2k and okodm modules, which already more or less have this setup.

This upgrade will mean an API break with previous versions. Be warned.