Added BarGraph widget
BarGraph widget can be used to display bargraphs. It also shows the min and
maximum values and a list with all the values so the user can copy them.
Furthermore there is a slot_update() function slot that can hold a pointer
to the function that can trigger a request for updated values. This slot
can be called using the refresh button or at regular intervals with the
auto-update button.