Bugfix: show vector ends in pixel center.
Displaying ref shifts and shifts goes as follows:
ref shift goes from subap center _exact_ to ref shift. However, ref shift
points to the origin of the pixel. We add (0.5, 0.5) to the end to show
the line to the center of the pixel.
Shift vec should start at a pixel center and end at a pixel center, so
add (0.5, 0.5) to both the start and the end.
Also modified in glviewer.