Implemted offset vector for WFC control
Becaues a wavefront corrector (such as a DM) is usually not flat when
actuated at 'zero', one can define an offset vector to be applied just
before actuating the mirror. This offset should be stored in Wfc::offset
and the final control vector Wfc::control, defined as
Wfc::ctrlparams.ctrl_vec - Wfc::offset, will be sent to the wavefront
corrector driver.