Do not reset() the mirror during influence calib
Do not set the mirror back to 'flat' (or whatever reset() does) in
between calculating the influence of each actuator: use the shape of the
mirror as set *before* calibrating the influence matrix.

If reset() gives a bad mirror shape, than we need to add some offset
before we can calculate a proper influence matrix. Setting the mirror
back with reset() would negate this offset every time.